pyproject.toml
[project]
name = "celline"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
authors = [
    { name = "YUYA556223", email = "yuya.sato.jp@gmail.com" }
]
requires-python = ">=3.10"
dependencies = [
    "argparse>=1.4.0",
    "continuousvi>=0.1.5",
    "inquirer>=3.4.0",
    "iprogress>=0.4",
    "ipywidgets>=8.1.5",
    "multipledispatch>=1.0.0",
    "pandas>=2.2.3",
    "polars>=1.26.0",
    "pyarrow>=19.0.1",
    "pyper>=1.1.2",
    "pysradb>=2.2.2",
    "pyyaml>=6.0.2",
    "requests-html>=0.10.0",
    "rich>=14.0.0",
    "scanpy>=1.11.1",
    "scrublet>=0.2.3",
    "toml>=0.10.2",
    "tqdm>=4.67.1",
    "varname>=0.14.0",
]

[project.scripts]
celline = "celline:main"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[dependency-groups]
dev = [
    "ipykernel>=6.29.5",
]

[tool.ruff]
extend-include = ["*.ipynb"]
exclude = [".venv/*", "*/resources/*", "*/data/*", "*/out/*", "*/dist/*"] # チェックを除外するディレクトリ
line-length = 200  # 1行の最大文字数
indent-width = 4  # インデント幅（スペースの数）

[tool.ruff.format]
docstring-code-format = true
indent-style = "space"  # または "tab"

[tool.ruff.lint]
select = ["ALL"]

ignore=[
    "ERA001", # commented out code
    "G004", # Logging
    "FBT001", # boolean type
    "FBT002", # boolean type
    "UP007", # Enable Optional[XXX]
    "EM101", #Exception
    "TRY003", #Exception
    "PLR0913", # long args
    "EM102", # f-string usage
    "D205", # Description
    "D401", # Doc string
    "N806", # Doc string
    "N999", # Public module description
    "C901", "PLR0915", # Args
    "PLR0912", # Branches
]

unfixable = [
    "F401",   # unused import
]
